refactor: to rebuild the project structure
