+++ image_hosting/utils/utils.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 //
 //func LskyRefreshToken() (string, error) {
